% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baron}
\alias{baron}
\title{Human pancreatic islet data}
\format{An object of class \code{list} of length 2.}
\usage{
baron
}
\description{
This is the Human pancreatic islet dataset (GSM2230757). The raw data contains 20,125 genes and 1,937
cells. Here we use the reference and downsampled datasets generated by Huang et al (2018) which contain
2,284 genes and 1,076 cells (available at https://github.com/mohuangx/SAVER-paper/tree/master/SAVER-data).
For details about the approach to generate the reference and downsampled
datasets, please refer to Huang et al (2018). This data is an object of class list of length two. count.ref is the
reference count matrix and count.samp is the downsampled count matrix.
}
\examples{
data("baron")
}
\references{
Baron, Maayan, et al (2016). A single-cell transcriptomic map of the human and mouse pancreas reveals inter-and intra-cell population structure. \emph{Cell systems}, 3(4):346-360.

Huang, M. et al. (2018). Saver: gene expression recovery for single-cell rna sequencing. Nat Methods, 15, 539–542.

Zhang, X. F. et al. (2019) EnImpute: imputing dropout events in single cell RNA sequencing data via ensemble learning.
}
\author{
Xiao-Fei zhang, \email{zhangxf@mail.ccnu.edu.cn}
}
\keyword{datasets}
