% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manno}
\alias{manno}
\title{Human ventral midbrain data}
\format{An object of class \code{list} of length 2.}
\usage{
manno
}
\description{
This is the Human ventral midbrain dataset (GSE76381). The raw data contains 19,531 genes and 1,977 cells.
Here we use the reference and downsampled datasets generated by Huang et al (2018) which contain
2,059 genes and 947 cells (available at https://github.com/mohuangx/SAVER-paper/tree/master/SAVER-data).
For details about the approach to generate the reference and downsampled
datasets, please refer to Huang et al (2018). This data is an object of class list of length two. count.ref is the
reference count matrix and count.samp is the downsampled count matrix.
}
\examples{
data("manno")
}
\references{
La Manno, Gioele, et al  (2016). Molecular diversity of midbrain development in mouse, human, and stem cells. Cell 167(2): 566-580.

Mo Huang et al (2018). SAVER: gene expression recovery for single-cell RNA sequencing.
 \emph{Nature Methods}, 15:539-542.
}
\author{
Xiao-Fei zhang, \email{zhangxf@mail.ccnu.edu.cn}
}
\keyword{datasets}
