% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zeisel}
\alias{zeisel}
\title{Mouse brain single-cell RNA-seq dataset}
\format{An object of class \code{list} of length 2.}
\usage{
zeisel
}
\description{
This is the mouse cortex and hippocampus dataset (http://linnarssonlab.org/cortex/). The raw data contains 19,972
genes and 3,005 cells. Here we use the reference and downsampled datasets generated by Huang et al (2018) which contain
3,529 genes and 1,799 cells (available at https://github.com/mohuangx/SAVER-paper/tree/master/SAVER-data). For details about the approach to generate the reference and downsampled
datasets, please refer to Huang et al (2018). This data is an object of class list of length two. count.ref is the
reference count matrix and count.samp is the downsampled count matrix.
}
\examples{
data("zeisel")
}
\references{
Amit Zeisel et al. (2015). Cell
types in the mouse cortex and hippocampus revealed by single-cell RNA-seq.
\emph{Science} 347(6226):1138-1142.

Mo Huang et al (2018). SAVER: gene expression recovery for single-cell RNA sequencing.
 \emph{Nature Methods}, 15:539-542.

Zhang, X. F. et al. (2019) EnImpute: imputing dropout events in single cell RNA sequencing data via ensemble learning.
}
\author{
Xiao-Fei zhang, \email{zhangxf@mail.ccnu.edu.cn}
}
\keyword{datasets}
